requirements.txt
aiohttp==3.11.7
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
cryptography==44.0.0
openai==1.55.0
langchain==0.3.8
langchain-openai==0.2.9
supabase==2.10.0
requests==2.32.3
Werkzeug==3.1.3
WTForms==3.2.1
email_validator==2.2.0
Crawl4AI==0.3.746
typing-inspect==0.9.0
typing_extensions==4.12.2
pydantic==2.10.1
pydantic_core==2.27.1
urllib3==2.2.3
gunicorn==21.2.0
asgiref==3.7.2
playwright==1.49.0
pyee==12.0.0
greenlet==3.1.1
celery==5.4.0
redis==5.2.1
flower==2.0.1
sqlalchemy_celery_beat==0.8.0
twilio == 9.4.4
emoji == 2.14.1
